refactor: payment UI
