+++ frontend/src/features/public-form/components/FormPaymentPage/components/FixedPaymentItemDetailsBlock.tsx
@@ -5,7 +5,7 @@ import { centsToDollars } from '~utils/payments'
 import PaymentItemNameDescription from './PaymentItemNameDescription'
 import { FixedItemDetailProps } from './types'
 
-export const PaymentItemDetailsBlock = ({
+export const FixedPaymentItemDetailsBlock = ({
   paymentItemName,
   paymentDescription,
   colorTheme,
