+++ frontend/src/features/public-form/components/FormPaymentPage/components/PaymentStack.tsx
@@ -1,6 +1,9 @@
 import React from 'react'
 import { Divider, Stack } from '@chakra-ui/react'
 
+/**
+ * Presentational wrapper over chakra-ui's `<Stack />` for payment elements.
+ */
 export const PaymentStack = ({ children }: { children: React.ReactNode }) => (
   <Stack
     spacing={{ base: '1.5rem', md: '2.25rem' }}
