+++ frontend/src/features/public-form/components/FormPaymentPage/components/assets/PaymentSuccessSvgr.tsx
