+++ frontend/src/features/public-form/components/FormPaymentPage/components/index.tsx
@@ -1,6 +1,6 @@
+export * from './assets/PaymentSuccessSvgr'
 export * from './CreatePaymentIntentFailureBlock'
-export * from './PaymentItemDetailsBlock'
+export * from './FixedPaymentItemDetailsBlock'
 export * from './PaymentStack'
-export * from './PaymentSuccessSvgr'
 export * from './ProductPaymentItemDetailsBlock'
 export * from './VariablePaymentItemDetailsBlock'
