+++ frontend/src/features/public-form/components/FormPaymentPage/stripe/components/index.tsx
@@ -1,3 +1,3 @@
 export * from './DownloadReceiptBlock'
 export * from './GenericMessageBlock'
-export * from './PaymentBlock'
+export * from './StripePaymentBlock'
